requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==41.0.3
Flask==2.3.2
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.1.0
pycparser==2.22
pycrypto==2.6.1
pycryptodome==3.22.0
pyotp==2.8.0
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
